+++ src/App.vue
@@ -3,14 +3,20 @@ SPDX-License-Identifier: Apache-2.0
 SPDX-FileCopyrightText: © 2025 Mercedes-Benz Tech Innovation GmbH
 -->
 <template>
-  <div class="h-full flex min-h-full border border-gray-300 dark:border-gray-200 rounded-md overflow-hidden">
+  <div
+    data-testid="ifex-viewer-container"
+    :class="currentSidenavPosition === 'left' ? 'flex-row' : 'flex-row-reverse'"
+    class="h-full flex min-h-full border border-gray-300 dark:border-gray-200 rounded-md overflow-hidden"
+  >
     <template v-if="activeView?.length">
       <Sidenav
         :selected-node-id="selectedNode?.id ?? ''"
         :tree-model="activeView"
         :show-tabs="hasMoreThanOneLayer"
+        :sidenav-position="currentSidenavPosition"
         @node-selected="onNodeSelected"
         @view-tab-changed="viewerModelStore.changeSelectedView"
+        @sidenav-position-changed="changeSidenavPosition"
       />
 
       <div class="flex justify-center w-full pt-6 px-8 bg-white dark:bg-gray-800 overflow-auto scrollbar-gutter">
@@ -45,28 +51,42 @@ SPDX-FileCopyrightText: © 2025 Mercedes-Benz Tech Innovation GmbH
 </template>
 
 <script setup lang="ts">
-import { computed, provide, watch } from 'vue';
+import { computed, provide, ref, watch } from 'vue';
 import { Breadcrumb, BreadcrumbsIconType } from './components/breadcrumbs/types';
 import Sidenav from './components/sidenav/Sidenav.vue';
-import { ClipboardCopiedEvent, CopiedSuccessfulEventKey, IfexViewerProps, NodeSelectedEvent } from './types';
+import { ClipboardCopiedEvent, CopiedSuccessfulEventKey, IfexViewerProps, NodeSelectedEvent, SidenavPosition } from './types';
 import { useDetailPageSelection } from './use-detail-page-selection';
-import { findNodeByPath, getFullPathToNode } from './utils/tree/tree';
+import { findNodeByPath, getFullDotNotationPath, getFullPathToNode } from './utils/tree/tree';
 import Breadcrumbs from './components/breadcrumbs/Breadcrumbs.vue';
 import { ExclamationTriangleIcon } from '@heroicons/vue/20/solid';
 import { useComplexDatatypesStore } from './stores/complex-datatypes/complex-datatypes.store.ts';
 import { useViewerModelStore } from './stores/viewer-model/viewer-model.store.ts';
 import { storeToRefs } from 'pinia';
 
-const { specifications } = defineProps<IfexViewerProps>();
+const {
+  specifications,
+  layout = {
+    sidenavPosition: 'left',
+  },
+} = defineProps<IfexViewerProps>();
 
 const viewerModelStore = useViewerModelStore();
 const { viewerModel, activeView, viewerModelWithoutApi } = storeToRefs(viewerModelStore);
+const { changeSelection, detailPage, selectCoreLayer, selectedNode } = useDetailPageSelection(activeView, viewerModelWithoutApi);
+
+// Keep track of the previously active node in order to restore the selection when the viewerModel changes (e.g. new specifications are loaded)
+const previouslyActiveNode = ref<string | null>(null);
 
 watch(
   () => specifications,
   (newSpecifications, oldSpecifications) => {
     if (!newSpecifications) {
       return;
+    }
+
+    if (selectedNode.value?.id) {
+      const currentlySelectedNodeId = selectedNode.value.id;
+      previouslyActiveNode.value = getFullDotNotationPath(currentlySelectedNodeId, viewerModelWithoutApi.value);
     }
 
     const nonEmptySpecifications = newSpecifications.filter(spec => !!spec?.content);
@@ -94,7 +114,12 @@ watch(
   { immediate: true },
 );
 
-const emits = defineEmits<{ specloaded: []; clipboardcopiedsuccessful: [payload: ClipboardCopiedEvent]; nodeselected: [payload: NodeSelectedEvent] }>();
+const emits = defineEmits<{
+  specloaded: [];
+  clipboardcopiedsuccessful: [payload: ClipboardCopiedEvent];
+  nodeselected: [payload: NodeSelectedEvent];
+  sidenavPositionChanged: [payload: SidenavPosition];
+}>();
 
 const hasMoreThanOneLayer = computed(() => viewerModel.value.layeredView.length > 1);
 
@@ -109,8 +134,6 @@ const onNodeSelected = (nodeId: string) => {
     emits('nodeselected', { path: dotNotationPath });
   }
 };
-
-const { changeSelection, detailPage, selectCoreLayer, selectedNode } = useDetailPageSelection(activeView, viewerModelWithoutApi);
 
 const breadcrumbs = computed<Breadcrumb[]>(() => {
   const selectedNodeVal = selectedNode.value;
@@ -143,15 +166,40 @@ const getPublicSlotName = (dotNoationPath: string) => dotNoationPath.toLowerCase
 
 const complexDatatypeStore = useComplexDatatypesStore();
 
+const newSpecLoaded = () => {
+  previouslyActiveNode.value = null;
+  complexDatatypeStore.setComplexDatatypes(activeView.value);
+  emits('specloaded');
+};
+
 watch(
   activeView,
   () => {
-    selectCoreLayer();
-    complexDatatypeStore.setComplexDatatypes(activeView.value);
-    emits('specloaded');
+    if (!previouslyActiveNode.value) {
+      selectCoreLayer();
+      newSpecLoaded();
+      return;
+    }
+
+    const node = findNodeByPath(previouslyActiveNode.value, '', viewerModelWithoutApi.value);
+
+    if (node?.id) {
+      changeSelection(node.id);
+    } else {
+      selectCoreLayer();
+    }
+
+    newSpecLoaded();
   },
   { immediate: true },
 );
+
+const currentSidenavPosition = ref(layout.sidenavPosition);
+
+const changeSidenavPosition = () => {
+  currentSidenavPosition.value = currentSidenavPosition.value === 'left' ? 'right' : 'left';
+  emits('sidenavPositionChanged', currentSidenavPosition.value);
+};
 
 /**
  * Selects a node in the sidenav by the given path.