+++ src/types.ts
@@ -6,10 +6,31 @@
 import { InjectionKey } from 'vue';
 import { AST } from './types/ifex-core';
 
-export interface IfexViewerProps {
-  specifications: IfexSpecificationItem[];
+export type SidenavPosition = 'left' | 'right';
+
+export interface IfexViewerLayout {
+  /**
+   * Determines whether the side navigation is positioned on the left or right side of the viewer.
+   */
+  sidenavPosition: SidenavPosition;
 }
 
+export interface IfexViewerProps {
+  /**
+   * An array of IFEX specification items to be displayed in the viewer.
+   */
+  specifications: IfexSpecificationItem[];
+  /**
+   * Specifies the layout configuration for the IFEX viewer.
+   */
+  layout: IfexViewerLayout;
+}
+
+/**
+ * Represents a single IFEX specification file with its filename and content.
+ * - `filename`: The name of the specification file (e.g. your-spec.yml).
+ * - `content`: The content of the specification file in YAML format as a string.
+ */
 export interface IfexSpecificationItem {
   filename: string;
   content: string;